clean up all kinds of shit in the comment model and migration and also make commenting work.
So, the comment migration table and model were all fucked up.  Like, fucked fucked.  Doubleplusfucked.
That's why ajax was being a total bitch all the fucking time.  Ajax is a real shithead, but also, before anyone
ever declares their work done, please please please just check that shit.  The photos and comments weren't linked up and had a lot of bugs
that were really shitty to work through.  Now the comments work.  Not the ajax yet.  But it's cool.  We're not going to worry about it just yet.